@@ -1,19 +1,3 @@
-/*
-* Copyright (C) 2018 HansooLabs.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
-
 package com.hansoolabs.calendarproto;
 
 import android.content.Context;
@@ -39,6 +23,7 @@ import fr.castorflex.android.verticalviewpager.VerticalViewPager;
  * Fragment for displaying the monthly-calendar
  * @author brownsoo
  *
+ * @noinspection unused
  */
 public class MonthlyFragment extends Fragment implements MonthlyView {
 
@@ -70,7 +55,6 @@ public class MonthlyFragment extends Fragment implements MonthlyView {
 
     @Nullable
     private MonthlyFragmentListener listener = null;
-    // Vertical View Pager -- ref: https://github.com/LittlePanpc/VerticalViewPager-1
     private VerticalViewPager vvPager;
     private MonthlyPagerAdapter adapter;
     private int mYear = -1;
@@ -113,7 +97,7 @@ public class MonthlyFragment extends Fragment implements MonthlyView {
     }
 
     @Override
-    public void onAttach(Context context) {
+    public void onAttach(@NonNull Context context) {
         super.onAttach(context);
         adapter = new MonthlyPagerAdapter(context, this);
     }
