@@ -1,19 +1,3 @@
-/*
-* Copyright (C) 2018 HansooLabs.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
-
 package com.hansoolabs.calendarproto.cal;
 
 import java.util.Calendar;
@@ -26,6 +10,7 @@ import com.hansoolabs.calendarproto.cal.WeatherInfo.Weather;
  * Value object for a day
  * @author brownsoo
  *
+ * @noinspection unused
  */
 public class OneDayData {
     
@@ -96,7 +81,7 @@ public class OneDayData {
     
     /**
      * Get weather info
-     * @return
+     * @return Weather
      */
     public Weather getWeather() {
         return this.weather;
